third_party/blink/web_tests/external/wpt/editing/run/delete_1-1000-expected.txt

This is a testharness.js-based test.
Found 14 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["delete",""]] "<span>foo[</span><span>]bar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span>foo</span><span>bar</span>" but got "<span>fo</span><span>bar</span>"
[FAIL] [["stylewithcss","true"],["delete",""]] "foo<span style=display:none>bar</span>[]baz" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["delete",""]] "<a>foo</a>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobar" but got "<a>fo</a>bar"
[FAIL] [["delete",""]] "<a href=/>foo</a>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobar" but got "<a href=\\"/\\">fo</a>bar"
[FAIL] [["delete",""]] "<a name=abc>foo</a>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobar" but got "<a name=\\"abc\\">fo</a>bar"
[FAIL] [["delete",""]] "<a href=/ name=abc>foo</a>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobar" but got "<a href=\\"/\\" name=\\"abc\\">fo</a>bar"
[FAIL] [["delete",""]] "<span><a>foo</a></span>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span>foo</span>bar" but got "<span><a>fo</a></span>bar"
[FAIL] [["delete",""]] "<span><a href=/>foo</a></span>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span>foo</span>bar" but got "<span><a href=\\"/\\">fo</a></span>bar"
[FAIL] [["delete",""]] "<span><a name=abc>foo</a></span>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span>foo</span>bar" but got "<span><a name=\\"abc\\">fo</a></span>bar"
[FAIL] [["delete",""]] "<span><a href=/ name=abc>foo</a></span>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span>foo</span>bar" but got "<span><a href=\\"/\\" name=\\"abc\\">fo</a></span>bar"
[FAIL] [["delete",""]] "foo <span>&nbsp;</span>[] bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo&nbsp; bar" but got "foo &nbsp;bar"
[FAIL] [["delete",""]] "foo <span>&nbsp;</span> []bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo&nbsp;<span> </span>bar" but got "foo <span>&nbsp;</span>bar"
[FAIL] [["delete",""]] "<b>foo&nbsp;</b> []bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<b>foo </b>bar" but got "<b>foo&nbsp;</b>bar"
[FAIL] [["delete",""]] "<b>foo&nbsp;</b>&nbsp;[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<b>foo </b>bar" but got "<b>foo&nbsp;</b>bar"
Harness: the test ran to completion.
